feat(tracing): instrument Task.Call
When the worker processes a task it will try to extract the trace context
from the signature's headers and start a new span from it. If it can't find
one start a trace anyway.

The Task.Call function will finish the newly started span as soon
as the function returns. It will also tag a span with `error=true`
if the call fails along with a stack trace that can be viewed in the span.