Redis backend - reset group expiry in TriggerChord
This fixes a bug where the group key TTL is reset to -1 by the SET call. We need to reset it back to our desired expiry after the SET call.

A future improvement could be to pass the EX option to SET instead of always doing a separate EXPIREAT for all keys.